feat: latest styles on cards

+++ 08-CRUD/app/js/index.js
@@ -13,7 +13,7 @@ document.addEventListener("DOMContentLoaded", function () {
 
   if (!tasks) {
     const p = document.createElement("p");
-    const text = document.createTextNode("No hay elementos.");
+    const text = document.createTextNode("Add a new task.");
     p.appendChild(text);
     tasksContainer.appendChild(p);
   } else {
@@ -23,7 +23,7 @@ document.addEventListener("DOMContentLoaded", function () {
       }
     } else {
       const p = document.createElement("p");
-      const text = document.createTextNode("No hay elementos.");
+      const text = document.createTextNode("Add a new task.");
       p.appendChild(text);
       tasksContainer.appendChild(p);
     }
@@ -35,6 +35,7 @@ document.addEventListener("DOMContentLoaded", function () {
 
     if (newTask) {
       tasks.push(newTask);
+      tasks.reverse();
       localStorage.setItem("tasks", JSON.stringify(tasks));
 
       tasksContainer.innerHTML = "";
@@ -48,42 +49,53 @@ document.addEventListener("DOMContentLoaded", function () {
   });
 
   function showTasks(tasks, i) {
-    const div = document.createElement("div");
-    const p = document.createElement("p");
-    const text = document.createTextNode(tasks[i]);
+    if (tasks.length === 0) {
+      const p = document.createElement("p");
+      const text = document.createTextNode("Add a new task.");
+      p.appendChild(text);
+      tasksContainer.appendChild(p);
+      tasksContainer.style.overflowY = "hidden";
+    } else {
+      const div = document.createElement("div");
+      const controlPanel = document.createElement("aside");
+      const p = document.createElement("p");
+      const text = document.createTextNode(tasks[i]);
 
-    const button = document.createElement("button");
-    const textDel = document.createTextNode("Delete");
+      const button = document.createElement("button");
+      const textDel = document.createTextNode("Delete");
 
-    const upBtn = document.createElement("button");
-    const textUp = document.createTextNode("Edit");
-    p.appendChild(text);
+      const upBtn = document.createElement("button");
+      const textUp = document.createTextNode("Edit");
+      p.appendChild(text);
 
-    button.appendChild(textDel);
-    upBtn.appendChild(textUp);
+      button.appendChild(textDel);
+      upBtn.appendChild(textUp);
 
-    button.onclick = () => {
-      deleteItem(i);
-    };
+      button.onclick = () => {
+        deleteItem(i);
+      };
 
-    upBtn.onclick = () => {
-      if (!editing) {
-        update(i);
-      } else {
-        alert("You're editing a task right now.");
-      }
-    };
+      upBtn.onclick = () => {
+        if (!editing) {
+          update(i);
+        } else {
+          alert("You're editing a task right now.");
+        }
+      };
 
-    div.appendChild(p);
-    div.appendChild(button);
-    div.appendChild(upBtn);
+      div.appendChild(p);
+      div.appendChild(controlPanel);
+      controlPanel.appendChild(button);
+      controlPanel.appendChild(upBtn);
 
-    tasksContainer.appendChild(div);
+      tasksContainer.appendChild(div);
+    }
   }
 
   function deleteItem(item) {
     const tasks = JSON.parse(localStorage.getItem("tasks")) || [];
     tasks.splice(item, 1);
+    tasks.reverse();
 
     localStorage.setItem("tasks", JSON.stringify(tasks));
 
@@ -106,10 +118,13 @@ document.addEventListener("DOMContentLoaded", function () {
 
       submitBtn.disabled = true;
       taskInput.disabled = true;
+      taskInput.style.display = "none";
+      submitBtn.style.display = "none";
 
       confirmBtn.onclick = () => {
         if (newTask.value != "") {
           tasks[item] = newTask.value;
+          tasks.reverse();
 
           localStorage.setItem("tasks", JSON.stringify(tasks));
           tasksContainer.innerHTML = "";
@@ -121,6 +136,8 @@ document.addEventListener("DOMContentLoaded", function () {
           editing = false;
           newTask.value = "";
           menu.style.display = "none";
+          taskInput.style.display = "block";
+          submitBtn.style.display = "block";
           submitBtn.disabled = false;
           taskInput.disabled = false;
         } else {
